chore: Add link sorting logic to make links blend in a bit.
May eventually make this an option to where people can choose their sort order (built in way, alphabetical or custom) but for now I'm leaving it like this.

I'll add more links to be sorted as I find them, but for now if a link is missing from the `LINK_ORDER` const then it gets sorted to the end.