[WIP] Various changes and starting of moving stuff over to using my utils
There is still a lot to do it's just I've been up for nearly ~24 hours now so I'm dead tired and want to stop working on this for a moment.

I'll get back to this soon though.. I'm not bumping versions numbers or anything until I get ready to start pushing this over to the main branch.